backend/requirements.txt
fastapi
uvicorn
httpx
nepse-api
